Restructured /internal/core for improved clarity
